Refactor UpdateChecker classes
